Fix type issue in mailto parsing

+++ indra/llui/llurlentry.cpp
@@ -1430,7 +1430,7 @@ LLUrlEntryEmail::LLUrlEntryEmail()
 
 std::string LLUrlEntryEmail::getLabel(const std::string &url, const LLUrlLabelCallback &cb)
 {
-	int pos = url.find("mailto:");
+	size_t pos = url.find("mailto:");
 
 	if (pos == std::string::npos)
 	{
